Fixed bug in fix{Freeze,Melt}
Since glass does not reduce the light level the height map does not
give us the highest non-air block, if that block is glass, because the
height map saves the block with the highest Y value that could get full
sunlight.

This will slow down fixing the weather conditions, but now snow can not
fall through glass.